Rename plugin to httpjson Add documentation on README
